Rental-Django/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
drf-yasg==1.21.8
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.2
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
whitenoise==6.8.2
